perf(voice): extract VoiceMeter to isolate 30 Hz re-renders
Subscribing to `voice://level` from inside `ChatInputArea` was
triggering a re-render of the entire ~1k-line composer subtree on every
backend tick (~30/sec). Move the listener and `vuLevel` state into a
small `VoiceMeter` child so only the meter re-renders. The parent
keeps the `useDynamicMeter` decision (depends on `prefers-reduced-motion`
and the active provider's `recordingMode`) and passes it down as a prop.
